requirements.txt
chromadb==1.0.15
llama-index==0.12.52
llama-index-embeddings-huggingface==0.5.5
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.4.7
llama-index-readers-file==0.4.11
llama-index-retrievers-bm25==0.5.2
llama-index-vector-stores-chroma==0.4.2
pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
streamlit==1.47.1